Move QR and add tests
